# src/main/resources/analitics/openapi-summary.yaml
...
openapi: 3.0.3
info:
  title: VTB AI HR Assistant API
  description: Backend API for AI-powered HR interview automation platform
  version: 1.0.0
  contact:
    name: VTB AI HR Team
    email: support@vtbaihr.com
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api-staging.vtbaihr.com/api/v1
    description: Staging server
  - url: https://api.vtbaihr.com/api/v1
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VacancyDto:
      type: object
      required: [title, department, location, type, description]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        department: { type: string }
        location: { type: string }
        type: { $ref: '#/components/schemas/EmploymentType' }
        status: { $ref: '#/components/schemas/VacancyStatus' }
...
    VacancyStatus:
      type: string
      enum: [DRAFT, ACTIVE, PAUSED, CLOSED, ARCHIVED]
    EmploymentType:
      type: string
      enum: [FULL_TIME, PART_TIME, CONTRACT, INTERNSHIP, FREELANCE]
...
paths:
...
  /vacancies:
    get:
      tags: [Vacancies]
      summary: List vacancies
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/VacancyStatus'
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: List of vacancies
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/VacancyDto'
                  totalElements: { type: integer }
                  totalPages: { type: integer }
    
    post:
      tags: [Vacancies]
      summary: Create vacancy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVacancyRequest'
      responses:
        '201':
          description: Vacancy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacancyDto'
  /vacancies/{id}:
    get:
      tags: [Vacancies]
      summary: Get vacancy by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Vacancy details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacancyDto'
    
    put:
      tags: [Vacancies]
      summary: Update vacancy
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVacancyRequest'
      responses:
        '200':
          description: Vacancy updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacancyDto'
    
    delete:
      tags: [Vacancies]
      summary: Delete vacancy
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204':
          description: Vacancy deleted
  
  # Candidate Endpoints - See api-hr.md for details
...